+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Pending player invites now show in the member list of the party
 * Added ally system to parties, closes #28
 * Added a button to the admin party screen to open the claim chunks gui in admin mode, closes #31
+* Added a command to change the chunk amount for all parties, closes #14
 
 # 1.0.10